+++ Kernel/CMakeLists.txt
@@ -128,6 +128,7 @@ set(KERNEL_SOURCES
     Devices/Storage/SD/SDHostController.cpp
     Devices/Storage/SD/SDMemoryCard.cpp
     Devices/Storage/USB/BulkSCSIInterface.cpp
+    Devices/Storage/USB/BulkSCSIStorageDevice.cpp
     Devices/Storage/VirtIO/VirtIOBlockController.cpp
     Devices/Storage/VirtIO/VirtIOBlockDevice.cpp
     Devices/Storage/StorageController.cpp
